It looks like you provided the output of a diff command, which compares two sets of text. This is not a problem that can be solved with a specific answer.
However, I can try to summarize what changed in the code. It appears
that many autoloads were replaced with `require` statements, and some
key bindings were removed or modified.

If you would like me to help you understand these changes further,
please provide more context about what you are trying to accomplish,
such as what you are using this code for and what language it is
in:(it seems to be a mix of Emacs Lisp and other languages).